Big PR to address some of my issues.
This is stil a work in progress to some degree.
  - Missing some idempotennce around bootenvs is the biggee.

This PR tracks content separately and doesn't partially load it.
It uses the versioned upstream community content.

It also creates and admin user and removes rocketskates.

It also adds a clean function, two playbook files, and
some example running examples.